Less scaffolding
