--- containers/moose/src/typeDefs/query.js
@@ -1,11 +1,0 @@
-const { gql } = require("apollo-server");
-
-const query = gql`
-  type Query {
-    user(username: String!): User,
-  }
-`;
-
-module.exports = {
-  query,
-};
