--- containers/moose/src/typeDefs/types/index.js
@@ -1,5 +1,0 @@
-const { userType } = require("./userType");
-
-module.exports = {
-  userType,
-};
