--- containers/moose/src/typeDefs/types/userType.js
@@ -1,14 +1,0 @@
-const { gql } = require("apollo-server");
-
-const userType = gql`
-  type User {
-    id: ID!
-    name: String
-    username: String!
-    email: String
-  }
-`;
-
-module.exports = {
-  userType,
-};
