+++ containers/moose/src/typeDefs/user.js
@@ -1,6 +1,17 @@
 const { gql } = require("apollo-server");
 
-const mutation = gql`
+const user = gql`
+  type User {
+    id: ID!
+    name: String
+    username: String!
+    email: String
+  }
+
+  type Query {
+    user(username: String!): User,
+  }
+
   type Mutation {
     newUser(username: String!, name: String, email: String): NewUserResponse!
   }
@@ -13,5 +24,5 @@ const mutation = gql`
 `;
 
 module.exports = {
-  mutation,
+  user,
 };
